Create useFetchPaginator hook to encapsulate pagination logic
