+++ src/components/TranslatedBookCreator/index.js
@@ -8,7 +8,7 @@ import BookRegistrationForm from "components/TranslatedBookCreator/BookRegistrat
 import BookRegistrationConfirm from "components/TranslatedBookCreator/BookRegistrationConfirm";
 
 import { Trans } from "react-i18next";
-import usePromise from "utils/hooks/usePromise";
+import useFetch from "utils/hooks/useFetch";
 import TranslatedBookController from "api/controllers/translation";
 
 const { Step } = Steps;
@@ -45,7 +45,7 @@ export default function TranslatedBookCreator() {
         setTranslatedBookPublications
     ] = useBook({ authors: [], publications: [] });
 
-    const { loading, data, fetch } = usePromise(
+    const { loading, data, fetch } = useFetch(
         TranslatedBookController.registerTranslation
     );
